refactor(api): reclassify shared leaf sinks as platform, flip telemetry to error
Same precedent as submission-status.ts: five files depend only on platform/
or external packages and are called from most domains, not domain business
logic of their own bucket -- genai.ts (the app's one place talking to Vertex
AI), moderation-metrics.ts, knowledge-metrics.ts, telemetry-health.ts, and
delivery-metrics.ts. Reclassified to platform in FILE_BUCKET and physically
git-mv'd into apps/api/src/platform/, matching submission-status.ts's own
precedent of keeping physical location in sync with logical bucket.

Also mapped the two catalog files that were missing FILE_BUCKET entries
(catalog-enricher.ts, embedding-service.ts -- both genuinely catalog, both
newer than the Wave A classification pass).

With those fixed, apps/api/src/telemetry/ has zero real (value-level)
cross-bucket imports as an importer -- the first bucket clean enough to
flip. Added ENFORCED_BUCKETS to module-boundary-map.mjs as the single
source of truth for which buckets are error-enforced, and wired it into
eslint.config.mjs so gamedev/module-boundary is 'error' for
apps/api/src/telemetry/** in the main `eslint .` pass (verified by
temporarily reintroducing a violation and confirming it fails). Every other
bucket stays warn-only via `npm run module-boundary` until its own PR.

apps/api/src warnings: 179 -> 158; crossBucket: 132 -> ~108 (many buckets
besides telemetry benefit from the platform reclassifications, not just
the one being flipped this round).

notifications was NOT flipped this round: its one remaining violation
(operator-alerts.ts calling creation/job-state.ts's pure state-transition
functions) is genuine domain-logic coupling, not a shared-primitive
misclassification -- left for a follow-up once it's clear whether that
belongs in packages/contract (N2, already underway elsewhere per
packages/contract/src/job-state.ts) or stays a real N1 boundary exception.
