+++ apps/api/src/creation/editor-drafts.ts
@@ -17,7 +17,7 @@ import { isLiveAgentRound } from './code-surface.js';
 import type { GamesStore } from '../delivery/games-store.js';
 import { MAX_EDITOR_DRAFT_BYTES, type Store, type SubmissionRecord } from '../platform/store.js';
 import type { ContentChecker } from '../platform/moderation.js';
-import { logModerationRejection } from '../telemetry/moderation-metrics.js';
+import { logModerationRejection } from '../platform/moderation-metrics.js';
 import { MAX_UTTERANCE_LENGTH, applyAssistPatches, assistEnabled, type EditorAssistant } from './editor-assist.js';
 import type { EditingGate } from './creation-limits.js';
