+++ apps/api/src/creation/scorecard.test.ts
@@ -2,7 +2,7 @@ import { beforeEach, describe, expect, it } from 'vitest';
 import { buildApp } from '../platform/app.js';
 import { InMemoryStore, type TelemetryEvent } from '../platform/store.js';
 import { buildScorecard, runScorecardSweep } from './scorecard.js';
-import { summarizeGameHealth } from '../telemetry/telemetry-health.js';
+import { summarizeGameHealth } from '../platform/telemetry-health.js';
 import type { InternalAuthVerifier } from '../platform/internal-auth.js';
 import { MIN_FEEDBACK_FOR_THEMES, type FeedbackTheme, type ThemeExtractor } from '../community/feedback-themes.js';
