+++ apps/api/src/delivery/source-delivery.test.ts
@@ -5,7 +5,7 @@ import type { KitFileStore, KitTree } from '../agent-surface/kit-files.js';
 import { KIT_ROOT_DIR } from '../agent-surface/kit-registry.js';
 import { InMemoryStore } from '../platform/store.js';
 import { NoopTranslator, type BilingualText, type Translator } from '../platform/translate.js';
-import { DELIVERY_ACCEPTED_MSG, DELIVERY_PREFLIGHT_REFUSED_MSG } from '../telemetry/delivery-metrics.js';
+import { DELIVERY_ACCEPTED_MSG, DELIVERY_PREFLIGHT_REFUSED_MSG } from '../platform/delivery-metrics.js';
 import {
   createSourceDeliveryService,
   SourceDeliveryAuthorityError,
