+++ apps/api/src/platform/genai.ts
