+++ apps/api/src/platform/knowledge-metrics.test.ts
