+++ apps/api/src/platform/knowledge-metrics.ts
