+++ eslint-rules/comment-prose-baseline.json
@@ -255,10 +255,13 @@
     "apps/api/src/platform/erase-account.ts": 64,
     "apps/api/src/platform/erase-player-signals.test.ts": 417,
     "apps/api/src/platform/erase-player-signals.ts": 1404,
+    "apps/api/src/platform/genai.ts": 148,
     "apps/api/src/platform/image-variants.test.ts": 98,
     "apps/api/src/platform/image-variants.ts": 261,
     "apps/api/src/platform/internal-auth.ts": 353,
     "apps/api/src/platform/localize-intake.ts": 586,
+    "apps/api/src/platform/moderation-metrics.test.ts": 610,
+    "apps/api/src/platform/moderation-metrics.ts": 489,
     "apps/api/src/platform/moderation-terms.ts": 236,
     "apps/api/src/platform/moderation.test.ts": 99,
     "apps/api/src/platform/moderation.ts": 336,
@@ -280,6 +283,8 @@
     "apps/api/src/platform/store.ts": 817,
     "apps/api/src/platform/submission-status.test.ts": 35,
     "apps/api/src/platform/submission-status.ts": 257,
+    "apps/api/src/platform/telemetry-health.test.ts": 583,
+    "apps/api/src/platform/telemetry-health.ts": 1074,
     "apps/api/src/platform/translate.test.ts": 108,
     "apps/api/src/platform/translate.ts": 1120,
     "apps/api/src/platform/verify-erase-signals.test.ts": 431,
