+++ eslint-rules/module-size-baseline.json
@@ -369,6 +369,8 @@
     "apps/api/src/platform/canonical-base64.ts": 36,
     "apps/api/src/platform/credential-scan.test.ts": 66,
     "apps/api/src/platform/credential-scan.ts": 38,
+    "apps/api/src/platform/delivery-metrics.test.ts": 117,
+    "apps/api/src/platform/delivery-metrics.ts": 126,
     "apps/api/src/platform/dev-seed-studio.ts": 195,
     "apps/api/src/platform/digest.test.ts": 315,
     "apps/api/src/platform/digest.ts": 310,
@@ -376,11 +378,16 @@
     "apps/api/src/platform/erase-account.ts": 49,
     "apps/api/src/platform/erase-player-signals.test.ts": 255,
     "apps/api/src/platform/erase-player-signals.ts": 234,
+    "apps/api/src/platform/genai.ts": 47,
     "apps/api/src/platform/image-variants.test.ts": 91,
     "apps/api/src/platform/image-variants.ts": 91,
     "apps/api/src/platform/internal-auth.test.ts": 95,
     "apps/api/src/platform/internal-auth.ts": 98,
+    "apps/api/src/platform/knowledge-metrics.test.ts": 60,
+    "apps/api/src/platform/knowledge-metrics.ts": 23,
     "apps/api/src/platform/localize-intake.ts": 95,
+    "apps/api/src/platform/moderation-metrics.test.ts": 180,
+    "apps/api/src/platform/moderation-metrics.ts": 86,
     "apps/api/src/platform/moderation-terms.ts": 120,
     "apps/api/src/platform/moderation.test.ts": 228,
     "apps/api/src/platform/moderation.ts": 247,
@@ -408,6 +415,8 @@
     "apps/api/src/platform/submission-token.ts": 52,
     "apps/api/src/platform/sweep-scope.test.ts": 33,
     "apps/api/src/platform/sweep-scope.ts": 18,
+    "apps/api/src/platform/telemetry-health.test.ts": 482,
+    "apps/api/src/platform/telemetry-health.ts": 363,
     "apps/api/src/platform/theme-css-generator.ts": 32,
     "apps/api/src/platform/translate.test.ts": 170,
     "apps/api/src/platform/translate.ts": 272,
